Revert "Split train"
