+++ .gitignore
@@ -10,4 +10,3 @@ Debug_Wayfinding
 .venv
 *.onnx
 checkpoints
-*.wbproj
