+++ src/Simulateur/worlds/.piste.wbproj
@@ -1,0 +1,9 @@
+Webots Project File version R2023b
+perspectives: 000000ff00000000fd0000000200000001000001cf00000278fc0200000001fb0000001400540065007800740045006400690074006f00720100000016000002780000008900ffffff000000030000078000000176fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000006900ffffff000005af0000027800000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff0000000100000002000001520000045b0100000002010000000100
+sceneTreePerspectives: 000000ff00000001000000030000001f000000c0000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: 2 "Bobox.proto" "../../CoVAPSy_Intech/Simulateur/worlds/piste.wbt" "../../CoVAPSy_Intech/Simulateur/protos/Vehicle.proto"
+consoles: Console:All:All
