+++ src/Simulateur/worlds/.piste0.wbproj
@@ -1,0 +1,11 @@
+Webots Project File version R2025a
+perspectives: 000000ff00000000fd00000002000000010000011c00000177fc0200000001fb0000001400540065007800740045006400690074006f00720000000016000001770000004500ffffff00000003000006c000000216fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000006c00000008700ffffff000006c0000001c400000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff0000000100000002000001000000017d0100000002010000000100
+sceneTreePerspectives: 000000ff00000001000000030000001f0000013e000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+globalOptionalRendering: LidarRaysPaths::LidarPointClouds
+consoles: Console:All:All
+renderingDevicePerspectives: TT02_0:RASPI_Camera_V2;1;32;0;0
