+++ src/Simulateur/worlds/.piste1.wbproj
@@ -1,0 +1,13 @@
+Webots Project File version R2025a
+perspectives: 000000ff00000000fd0000000200000001000000870000028afc0200000001fb0000001400540065007800740045006400690074006f007201000000000000028a0000004500ffffff00000003000006c000000150fc0100000002fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000006c00000008700fffffffb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c010000000000000a000000000000000000000006370000028a00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff00000001000000020000012b000005a50100000002010000000100
+sceneTreePerspectives: 000000ff00000001000000030000001e00000364000000fa0100000002010000000200
+minimizedPerspectives: 000000ff00000000fd0000000200000001000000750000017bfc0200000001fb0000001400540065007800740045006400690074006f007201000000160000017b0000003f00ffffff000000030000039b00000039fc0100000002fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c01000000000000039b0000006900fffffffb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c010000000000000a000000000000000000000003240000017b00000001000000020000000100000008fc00000000
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+globalOptionalRendering: LidarPointClouds::LidarRaysPaths
+consoles: Console:All:All
+renderingDevicePerspectives: TT02_0:RASPI_Camera_V2;1;32;0;0
+renderingDevicePerspectives: sparringpartner_car_0:RASPI_Camera_V2;1;32;0;0
