+++ src/Simulateur/worlds/.piste2.wbproj
@@ -1,0 +1,12 @@
+Webots Project File version R2025a
+perspectives: 000000ff00000000fd0000000200000001000001910000036ffc0200000001fb0000001400540065007800740045006400690074006f007200000000160000036f0000003f00ffffff000000030000039b00000039fc0100000002fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c01000000000000039b0000006900fffffffb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c01000000000000078000000000000000000000039b0000036a00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff00000001000000020000014d000006310100000002010000000100
+sceneTreePerspectives: 000000ff00000001000000030000001d0000013c000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+globalOptionalRendering: LidarPointClouds::LidarRaysPaths::ContactPoints::CameraFrustums
+consoles: Console:All:All
+renderingDevicePerspectives: TT02_0:RASPI_Camera_V2;1;32;0;0
+renderingDevicePerspectives: TT02_0:raspi_camera;1;1;0;0
